+++ EZOGroupFrames.lua
@@ -152,7 +152,7 @@ function ADDON.RegisterWithEZOCore()
             id = "ezogroupframes",
             name = ADDON.ADDON_NAME or "EZOGroupFrames",
             version = ADDON.ADDON_VERSION or "0.0.0",
-            addOnVersion = 116,
+            addOnVersion = 117,
             apiVersion = 1,
             capabilities = {
                 "family.language.consumer",